update the file structure to separate as much of the game logic from the main.py file as possible
